+++ docs/linkedlist/PartitionLinkedList.md
@@ -1,41 +1,15 @@
-<!----- Conversion time: 0.601 seconds.
-
-
-Using this Markdown file:
-
-1. Cut and paste this output into your source file.
-2. See the notes and action items below regarding this conversion run.
-3. Check the rendered output (headings, lists, code blocks, tables) for proper
-   formatting and use a linkchecker before you publish this page.
-
-Conversion notes:
-
-* GD2md-html version 1.0β13
-* Tue Jan 15 2019 20:00:27 GMT-0800 (PST)
-* Source doc: https://docs.google.com/open?id=1qyCkg6YXgTdMvgCzp1F1mDlndkZcAXKi5Co0VsdSngk
------>
-
-
-
 # Partition Linked List
 
 [https://app.laicode.io/app/problem/42](https://app.laicode.io/app/problem/42)
 
-
 ## Description
-
 
     Given a linked list and a target value T, partition it such that all nodes less than T are listed before the nodes larger than or equal to target value T. The original relative order of the nodes in each of the two partitions should be preserved.
 
 
     Examples
 
-
-
-*   L = 2 -> 4 -> 3 -> 5 -> 1 -> null, T = 3, is partitioned to 2 -> 1 -> 4 -> 3 -> 5 -> null
-
-
-
+- L = 2 -> 4 -> 3 -> 5 -> 1 -> null, T = 3, is partitioned to 2 -> 1 -> 4 -> 3 -> 5 -> null
 
 Medium
 
@@ -43,11 +17,9 @@ Linked List
 
 Partition
 
-
 ## Assumption
 
 The input linked list should not be null or empty.
-
 
 ## Algorithm
 
@@ -55,25 +27,18 @@ This question is similar to [QuickSort](../../../../A/SortingAlgorithms/Medium/Q
 
 We can do it in the following steps:
 
-
-
 1.  Create/allocate two new linked lists (dummy head)
-1.  Compare every node in the list with the target, put all the nodes that have values smaller than the target to newListOne. Otherwise, put them to newListTwo
+2.  Compare every node in the list with the target, put all the nodes that have values smaller than the target to newListOne. Otherwise, put them to newListTwo
     1.  If current.value < target: newListOne.append(current)
-    1.  else: newListTwo.append(current)
-1.  Concatenate newListOne and newListTwo and cut off at the end
+    2.  else: newListTwo.append(current)
+3.  Concatenate newListOne and newListTwo and cut off at the end
     1.  listTwoTail.next = null;
 
 Special attention should be paid to prevent a cycle from forming during the process
 
-
-
-
 ## Solution
 
-
 ### Code
-
 
 ```java
 /**
@@ -87,6 +52,7 @@ Special attention should be paid to prevent a cycle from forming during the proc
  * }
  */
 public class Solution {
+
   public ListNode partition(ListNode head, int target) {
     // Write your solution here
     if (head == null || head.next == null) {
@@ -120,13 +86,8 @@ public class Solution {
 }
 ```
 
-
-
 ### Complexity
 
 Time: there are n nodes to compare ⇒ O(n).
 
 Space: only dummy nodes are created, nothing lese ⇒ O(1).
-
-
-<!-- GD2md-html version 1.0β13 -->
